+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Version 5.0.1 (Under development)
 
+### App Center Distribute
+
+* **[Fix]** Fix NPE in Distribute module on resume distribute workflow when showing update setup failed dialog.
  ___
 
 ## Version 5.0.0