+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### AppCenter
 
 * **[Fix]** Fix Concurrent Modification Exception in DefaultChannel on setNetworkRequestsAllowed.
+* **[Improvement]** Migrate from Maven Dcendents to Maven Publish plugin.
 
 ### App Center Distribute